feat: fold the marketing site into apps/web
apps/landing is gone; apps/web now serves the marketing page at / and
the bb connect dashboard at /dashboard, with the landing worker's
/download/macos redirect and /api/subscribe endpoints ported to
TanStack server routes (framework-free logic + tests in
src/landing/endpoints.ts). Marketing CSS and the dashboard's Tailwind
theme both define :root tokens, so each loads via its own route-level
stylesheet link and cross-navigation stays full-page. turbo build
override, repository overview, and Resend/PostHog env plumbing updated.
