--- apps/landing/.gitignore
@@ -1,1 +1,0 @@
-.wrangler
