--- apps/landing/src/routes/index.tsx
@@ -1,1776 +1,0 @@
-import {
-  ArrowDown01Icon,
-  ArrowExpand01Icon,
-  ArrowLeft01Icon,
-  ArrowMoveDownLeftIcon,
-  ArrowRight01Icon,
-  AttachmentIcon,
-  BubbleChatAddIcon,
-  CheckmarkCircle02Icon,
-  Clock01Icon,
-  FolderGitTwoIcon,
-  FolderIcon as HiFolderIcon,
-  GitBranchIcon as HiGitBranchIcon,
-  GitMergeIcon as HiGitMergeIcon,
-  LaptopIcon as HiLaptopIcon,
-  Loading03Icon,
-  MessageQuestionIcon,
-  Mic02Icon,
-  MoreHorizontalIcon,
-  PlusMinusSquare01Icon,
-  SentIcon,
-  Settings01Icon,
-  SidebarLeftIcon,
-  SidebarRightIcon,
-  Tick02Icon,
-} from "@hugeicons/core-free-icons";
-import { HugeiconsIcon, type IconSvgElement } from "@hugeicons/react";
-import { createFileRoute } from "@tanstack/react-router";
-import { useCallback, useEffect, useMemo, useState } from "react";
-import type { CSSProperties, FormEvent, ReactNode } from "react";
-
-import { trackLandingEvent } from "../analytics";
-import bbIcon from "../assets/bb-icon.png";
-import hermesAvatar from "../assets/hermes-avatar.jpg";
-import vscodeIcon from "../assets/vscode.png";
-import { ClaudeIcon, CursorIcon, OpenAiIcon, OpencodeIcon, PiIcon } from "../icons";
-import type { CtaPlacement } from "../site";
-import { CLI_COMMAND, GITHUB_URL, SUBSCRIBE_PATH, downloadMacosHref } from "../site";
-
-export const Route = createFileRoute("/")({
-  component: LandingPage,
-});
-
-// Filled (solid) variant of the Hugeicons apple — same silhouette as their
-// stroke AppleIcon, but rendered as a fill so the macOS CTA reads as a solid
-// glyph (the free icon set ships outline variants only).
-const AppleSolidIcon: IconSvgElement = [
-  [
-    "path",
-    {
-      d: "M12 5.75C12 3.75 13.5 1.75 15.5 1.75C15.5 3.75 14 5.75 12 5.75Z",
-      fill: "currentColor",
-      key: "0",
-    },
-  ],
-  [
-    "path",
-    {
-      d: "M12.5 8.09001C11.9851 8.09001 11.5867 7.92646 11.1414 7.74368C10.5776 7.51225 9.93875 7.25 8.89334 7.25C7.02235 7.25 4 8.74945 4 12.7495C4 17.4016 7.10471 22.25 9.10471 22.25C9.77426 22.25 10.3775 21.9871 10.954 21.7359C11.4815 21.5059 11.9868 21.2857 12.5 21.2857C13.0132 21.2857 13.5185 21.5059 14.046 21.7359C14.6225 21.9871 15.2257 22.25 15.8953 22.25C17.2879 22.25 18.9573 19.8992 20 16.9008C18.3793 16.2202 17.338 14.618 17.338 12.75C17.338 11.121 18.2036 10.0398 19.5 9.25C18.5 7.75 17.0134 7.25 15.9447 7.25C14.8993 7.25 14.2604 7.51225 13.6966 7.74368C13.2514 7.92646 13.0149 8.09001 12.5 8.09001Z",
-      fill: "currentColor",
-      key: "1",
-    },
-  ],
-];
-
-/* ── CTAs ─────────────────────────────────────────────────────────── */
-
-type CtaLinkProps = {
-  placement: CtaPlacement;
-  /** Omit for a plain inline link (nav/footer); set for button-styled CTAs. */
-  className?: string;
-  children: ReactNode;
-};
-
-function DownloadLink({ placement, className, children }: CtaLinkProps) {
-  return (
-    <a className={className} href={downloadMacosHref(placement)}>
-      {children}
-    </a>
-  );
-}
-
-function GitHubLink({ placement, className, children }: CtaLinkProps) {
-  return (
-    <a
-      className={className}
-      href={GITHUB_URL}
-      target="_blank"
-      rel="noreferrer"
-      onClick={() =>
-        trackLandingEvent({
-          name: "landing_github_clicked",
-          properties: { placement },
-        })
-      }
-    >
-      {children}
-    </a>
-  );
-}
-
-// The browser install path, rendered as an outline button whose body is the
-// run command. Clicking anywhere copies it (there's no hosted URL to open —
-// the command starts bb locally and opens it in the browser).
-function RunCommandButton({ placement }: { placement: CtaPlacement }) {
-  const [copied, setCopied] = useState(false);
-  const copy = () => {
-    // Track and show feedback first; the clipboard write can reject (no user
-    // activation, permissions) and must not swallow the event.
-    trackLandingEvent({
-      name: "landing_cli_command_copied",
-      properties: { placement, command: CLI_COMMAND },
-    });
-    setCopied(true);
-    setTimeout(() => setCopied(false), 1500);
-    navigator.clipboard.writeText(CLI_COMMAND).catch(() => {});
-  };
-  return (
-    <button
-      type="button"
-      className={
-        copied
-          ? "btn btn-ghost btn-install cmd-btn copied"
-          : "btn btn-ghost btn-install cmd-btn"
-      }
-      onClick={copy}
-      aria-label={`Copy browser install command: ${CLI_COMMAND}`}
-    >
-      <span className="cmd-dollar">$</span>
-      <span className="cmd-text">{CLI_COMMAND}</span>
-      <span className="cmd-copy">Copy</span>
-      {/* Toast floats above the button (absolute) so confirming the copy never
-          reflows the centered CTA row — the label stays a fixed width. */}
-      <span
-        className={copied ? "cmd-toast show" : "cmd-toast"}
-        aria-hidden="true"
-      >
-        Copied to clipboard
-      </span>
-    </button>
-  );
-}
-
-function InstallOptions({ placement }: { placement: CtaPlacement }) {
-  return (
-    <div className="install-options">
-      <div className="install-actions">
-        <span className="install-choice">
-          <DownloadLink
-            placement={placement}
-            className="btn btn-primary btn-install"
-          >
-            <HugeiconsIcon icon={AppleSolidIcon} className="btn-ic" />
-            Download for macOS
-          </DownloadLink>
-          <span className="install-note">One-click, no terminal</span>
-        </span>
-        <span className="install-choice">
-          <RunCommandButton placement={placement} />
-          <span className="install-note">
-            Windows (via WSL), Linux &amp; remote machines
-          </span>
-        </span>
-      </div>
-    </div>
-  );
-}
-
-/* ── Email signup ─────────────────────────────────────────────────── */
-
-type SubscribeStatus = "idle" | "submitting" | "success" | "error";
-
-// Email capture that POSTs to the first-party /api/subscribe Worker route,
-// which adds the address to the bb marketing audience in Resend. JS-enhanced:
-// it submits inline and swaps to a confirmation rather than navigating.
-function EmailSignup({ placement }: { placement: CtaPlacement }) {
-  const [email, setEmail] = useState("");
-  const [status, setStatus] = useState<SubscribeStatus>("idle");
-  const [error, setError] = useState("");
-
-  const submit = async (event: FormEvent<HTMLFormElement>) => {
-    event.preventDefault();
-    if (status === "submitting") {
-      return;
-    }
-    setStatus("submitting");
-    setError("");
-    try {
-      const response = await fetch(SUBSCRIBE_PATH, {
-        body: JSON.stringify({ email }),
-        headers: { "content-type": "application/json" },
-        method: "POST",
-      });
-      if (!response.ok) {
-        const body = (await response.json().catch(() => ({}))) as {
-          error?: string;
-        };
-        setError(body.error ?? "Something went wrong. Try again.");
-        setStatus("error");
-        return;
-      }
-      trackLandingEvent({ name: "landing_email_subscribed", properties: { placement } });
-      setStatus("success");
-    } catch {
-      setError("Could not reach the server. Try again.");
-      setStatus("error");
-    }
-  };
-
-  if (status === "success") {
-    return (
-      <p className="subscribe-done" role="status">
-        <HugeiconsIcon icon={CheckmarkCircle02Icon} className="subscribe-done-ic" />
-        You&rsquo;re on the list. We&rsquo;ll be in touch.
-      </p>
-    );
-  }
-
-  return (
-    <form className="subscribe-form" onSubmit={submit} noValidate>
-      <input
-        className="subscribe-input"
-        type="email"
-        name="email"
-        inputMode="email"
-        autoComplete="email"
-        required
-        placeholder="you@example.com"
-        aria-label="Email address"
-        aria-invalid={status === "error"}
-        value={email}
-        onChange={(event) => {
-          setEmail(event.target.value);
-          if (status === "error") {
-            setStatus("idle");
-          }
-        }}
-      />
-      <button
-        type="submit"
-        className="btn btn-primary subscribe-btn"
-        disabled={status === "submitting"}
-      >
-        {status === "submitting" ? "Subscribing…" : "Subscribe"}
-      </button>
-      {status === "error" ? (
-        <span className="subscribe-error" role="alert">
-          {error}
-        </span>
-      ) : null}
-    </form>
-  );
-}
-
-/* ── Scroll reveal ────────────────────────────────────────────────── */
-
-/** Fade-up sections as they scroll into view. No-JS and prerender stay fully visible. */
-function useScrollReveal() {
-  useEffect(() => {
-    if (window.matchMedia("(prefers-reduced-motion: reduce)").matches) {
-      return;
-    }
-    const targets = Array.from(document.querySelectorAll("[data-reveal]"));
-    for (const target of targets) {
-      if (target.getBoundingClientRect().top > window.innerHeight * 0.9) {
-        target.classList.add("reveal-pending");
-      }
-    }
-    const observer = new IntersectionObserver(
-      (entries) => {
-        for (const entry of entries) {
-          if (entry.isIntersecting) {
-            entry.target.classList.remove("reveal-pending");
-            observer.unobserve(entry.target);
-          }
-        }
-      },
-      { rootMargin: "0px 0px -12% 0px" },
-    );
-    for (const target of targets) {
-      observer.observe(target);
-    }
-    return () => observer.disconnect();
-  }, []);
-}
-
-/** The app mock assembles itself the first time it scrolls into view: window
- *  frame, then title bar, sidebar rows, conversation, and composer in sequence.
- *  The mock is held hidden from first paint by CSS (`html.js` + `:not(.constructing)`)
- *  so it never flashes finished before it builds. Once the entrance finishes the
- *  class is swapped to `.constructed` so later re-renders (switching threads,
- *  opening the diff) don't replay it. Prerender/no-JS/reduced-motion render the
- *  finished mock with no animation. */
-function useConstructMock() {
-  useEffect(() => {
-    if (window.matchMedia("(prefers-reduced-motion: reduce)").matches) {
-      return;
-    }
-    const mock = document.querySelector("[data-construct]");
-    if (!mock || mock.classList.contains("constructed")) {
-      return;
-    }
-    let timer = 0;
-    const observer = new IntersectionObserver(
-      (entries) => {
-        for (const entry of entries) {
-          if (entry.isIntersecting) {
-            const el = entry.target;
-            el.classList.add("constructing");
-            observer.unobserve(el);
-            timer = window.setTimeout(() => {
-              el.classList.remove("constructing");
-              el.classList.add("constructed");
-            }, 1800);
-          }
-        }
-      },
-      // Threshold 0 (not a ratio) so a mock taller than a small mobile viewport
-      // still triggers; the bottom margin holds it until it is meaningfully in view.
-      { threshold: 0, rootMargin: "0px 0px -20% 0px" },
-    );
-    observer.observe(mock);
-    return () => {
-      observer.disconnect();
-      if (timer) {
-        window.clearTimeout(timer);
-      }
-    };
-  }, []);
-}
-
-/** Scale the desktop app mock for narrow viewports. Below the mobile breakpoint
- *  the mock keeps its full desktop layout and is shrunk with `zoom` so a fixed
- *  left slice of the app (`--mock-visible-width`) fills the available width; the
- *  rest bleeds off the right edge, clipped by `.mockup-wrap`'s overflow. This
- *  stays legible instead of shrinking the whole app to fit. `--mock-visible-width`
- *  is defined only inside that breakpoint, so above it the variable is unset and
- *  the mock renders unscaled at its natural width. */
-function useFitMock() {
-  useEffect(() => {
-    const mock = document.querySelector<HTMLElement>(".mock");
-    const wrap = mock?.parentElement;
-    if (!mock || !wrap) {
-      return;
-    }
-    const fit = () => {
-      const wrapStyle = getComputedStyle(wrap);
-      const visibleWidth = Number.parseFloat(
-        getComputedStyle(mock).getPropertyValue("--mock-visible-width"),
-      );
-      if (!visibleWidth) {
-        // Desktop layout (variable unset above the breakpoint): no scaling.
-        mock.style.removeProperty("--mock-scale");
-        return;
-      }
-      // The card is inset by the wrap's side padding (its left gutter holds the
-      // drop shadow), so its on-screen width is the content box — clientWidth
-      // minus the padding — not clientWidth itself.
-      const slice =
-        wrap.clientWidth -
-        Number.parseFloat(wrapStyle.paddingLeft) -
-        Number.parseFloat(wrapStyle.paddingRight);
-      mock.style.setProperty("--mock-scale", String(slice / visibleWidth));
-    };
-    fit();
-    const observer = new ResizeObserver(fit);
-    observer.observe(wrap);
-    return () => observer.disconnect();
-  }, []);
-}
-
-/* ── Shared bits ──────────────────────────────────────────────────── */
-
-function ProviderChips() {
-  return (
-    <>
-      <ClaudeIcon className="plogo" />
-      <OpenAiIcon className="plogo" />
-      <CursorIcon className="plogo" />
-      <PiIcon className="plogo" />
-      <OpencodeIcon className="plogo" />
-    </>
-  );
-}
-
-/* ── Hero: interactive bb app mock ────────────────────────────────── */
-// A faithful recreation of the bb app: icon rail + thread sidebar + a markdown
-// conversation + the real composer (PR/diff bar, model picker, worktree row).
-// Clicking a thread in the sidebar swaps the conversation and composer.
-
-type IconProps = { className?: string };
-
-// Real bb app icons (Hugeicons), matched to the app's own Icon map in
-// apps/app/src/components/ui/icon.tsx — same glyphs the desktop app renders.
-const PanelIcon = ({ className }: IconProps) => (
-  <HugeiconsIcon icon={SidebarLeftIcon} className={className} />
-);
-const PanelRightIcon = ({ className }: IconProps) => (
-  <HugeiconsIcon icon={SidebarRightIcon} className={className} />
-);
-const ChevronLeft = ({ className }: IconProps) => (
-  <HugeiconsIcon icon={ArrowLeft01Icon} className={className} />
-);
-const ChevronRight = ({ className }: IconProps) => (
-  <HugeiconsIcon icon={ArrowRight01Icon} className={className} />
-);
-const ChevronDown = ({ className }: IconProps) => (
-  <HugeiconsIcon icon={ArrowDown01Icon} className={className} />
-);
-const Ellipsis = ({ className }: IconProps) => (
-  <HugeiconsIcon icon={MoreHorizontalIcon} className={className} />
-);
-const NewThreadIcon = ({ className }: IconProps) => (
-  <HugeiconsIcon icon={BubbleChatAddIcon} className={className} />
-);
-const ClockIcon = ({ className }: IconProps) => (
-  <HugeiconsIcon icon={Clock01Icon} className={className} />
-);
-const GearIcon = ({ className }: IconProps) => (
-  <HugeiconsIcon icon={Settings01Icon} className={className} />
-);
-const CheckIcon = ({ className }: IconProps) => (
-  <HugeiconsIcon icon={Tick02Icon} className={className} />
-);
-// Sidebar thread-status glyphs, matching the real app's muted glyphs
-// (CheckmarkCircle02 for done, MessageQuestion for needs-input).
-const CircleCheckIcon = ({ className }: IconProps) => (
-  <HugeiconsIcon icon={CheckmarkCircle02Icon} className={className} />
-);
-const MessageQuestionGlyph = ({ className }: IconProps) => (
-  <HugeiconsIcon icon={MessageQuestionIcon} className={className} />
-);
-const PaperPlane = ({ className }: IconProps) => (
-  <HugeiconsIcon icon={SentIcon} className={className} />
-);
-const Paperclip = ({ className }: IconProps) => (
-  <HugeiconsIcon icon={AttachmentIcon} className={className} />
-);
-const FolderIcon = ({ className }: IconProps) => (
-  <HugeiconsIcon icon={HiFolderIcon} className={className} />
-);
-const FolderGitIcon = ({ className }: IconProps) => (
-  <HugeiconsIcon icon={FolderGitTwoIcon} className={className} />
-);
-const GitBranchIcon = ({ className }: IconProps) => (
-  <HugeiconsIcon icon={HiGitBranchIcon} className={className} />
-);
-const GitMergeIcon = ({ className }: IconProps) => (
-  <HugeiconsIcon icon={HiGitMergeIcon} className={className} />
-);
-const Spinner = ({ className }: IconProps) => (
-  <HugeiconsIcon icon={Loading03Icon} className={className} />
-);
-const Maximize2 = ({ className }: IconProps) => (
-  <HugeiconsIcon icon={ArrowExpand01Icon} className={className} />
-);
-const MicIcon = ({ className }: IconProps) => (
-  <HugeiconsIcon icon={Mic02Icon} className={className} />
-);
-const SendIcon = ({ className }: IconProps) => (
-  <HugeiconsIcon icon={ArrowMoveDownLeftIcon} className={className} />
-);
-const LaptopGlyph = ({ className }: IconProps) => (
-  <HugeiconsIcon icon={HiLaptopIcon} className={className} />
-);
-const FileDiffIcon = ({ className }: IconProps) => (
-  <HugeiconsIcon icon={PlusMinusSquare01Icon} className={className} />
-);
-
-type Status = "running" | "done" | "waiting";
-type Step =
-  | { kind: "user"; text: string }
-  | { kind: "step"; text: ReactNode }
-  | { kind: "say"; text: ReactNode }
-  // A "spawn" step prints a tool line in the feed and, the first time it
-  // streams in, adds a nested child thread to the sidebar (like the real app).
-  | { kind: "spawn"; text: ReactNode; child: MockThread };
-type Ask = {
-  question: string;
-  options: { label: string; description: string }[];
-  selected: number;
-};
-type MockThread = {
-  id: string;
-  title: string;
-  status: Status;
-  branch: string;
-  pr?: number;
-  change: { files: number; add: number; del: number };
-  transcript: Step[];
-  /** Endlessly-cycled work a running thread streams in after its transcript. */
-  stream?: Step[];
-  /** A pending AskUserQuestion that replaces the prompt box (like the app). */
-  ask?: Ask;
-};
-
-// The subagent the Sentry thread spawns mid-run. It lands as a nested child row
-// in the sidebar and, if opened, streams its own work like any running thread.
-const SENTRY_SUBAGENT: MockThread = {
-  id: "sentry-sub",
-  title: "Reproduce the null cart",
-  status: "running",
-  branch: "bb/triage-sentry-spike",
-  change: { files: 1, add: 14, del: 0 },
-  transcript: [
-    { kind: "user", text: "Reproduce the null cart in applyPromo." },
-    { kind: "step", text: "Read src/checkout/applyPromo.ts" },
-  ],
-  stream: [
-    { kind: "step", text: "Built an empty-cart fixture" },
-    {
-      kind: "say",
-      text: (
-        <>
-          An active promo on an empty <code>cart</code> throws. Reproduced.
-        </>
-      ),
-    },
-    { kind: "step", text: "Wrote a failing test" },
-    { kind: "say", text: "Handed the repro back to the parent thread." },
-    { kind: "step", text: "Re-checked the stack trace" },
-  ],
-};
-
-// Endless "work" each running thread streams in after its transcript. The pool
-// loops, so a glance at the hero always shows tool calls and messages arriving.
-const SENTRY_STREAM: Step[] = [
-  { kind: "step", text: "Ran 48 tests" },
-  {
-    kind: "say",
-    text: (
-      <>
-        All green. The null <code>cart</code> path is covered now.
-      </>
-    ),
-  },
-  {
-    kind: "spawn",
-    text: (
-      <>
-        Spawned a subagent: <strong>Reproduce the null cart</strong>
-      </>
-    ),
-    child: SENTRY_SUBAGENT,
-  },
-  { kind: "step", text: "Edited promo.test.ts" },
-  { kind: "say", text: "Added a case for an empty cart with an active promo." },
-  { kind: "step", text: "Checked Sentry for new events" },
-  { kind: "say", text: "No new occurrences in the last 10 minutes." },
-  { kind: "step", text: "Read applyPromo.ts" },
-  {
-    kind: "say",
-    text: (
-      <>
-        Tightening the type so <code>cart</code> can't be null at the call site.
-      </>
-    ),
-  },
-  { kind: "step", text: "Edited 2 files" },
-  { kind: "say", text: "Pushed the guard and a follow-up. Re-running the suite." },
-];
-
-const LIN482_STREAM: Step[] = [
-  { kind: "step", text: "Ran 12 tests" },
-  { kind: "say", text: "Debounce holds for 200ms. One call, asserted." },
-  {
-    kind: "step",
-    text: (
-      <>
-        Edited <code>SearchBar.tsx</code>
-      </>
-    ),
-  },
-  { kind: "say", text: "Cancelling the timer on unmount so there's no leak." },
-  { kind: "step", text: "Checked the other call sites" },
-  { kind: "say", text: "Two more inputs could reuse this. Noted it on LIN-482." },
-  { kind: "step", text: "Edited 1 file" },
-  { kind: "say", text: "Verifying the debounce once more." },
-];
-
-const CHIEF_STREAM: Step[] = [
-  { kind: "step", text: "Swept 4 active threads" },
-  { kind: "say", text: "Sentry triage is re-running tests; LIN-482 is verifying." },
-  { kind: "step", text: "Checked for blockers" },
-  {
-    kind: "say",
-    text: (
-      <>
-        One thread is waiting on you: <code>Refactor the timeline cache</code>.
-      </>
-    ),
-  },
-  { kind: "step", text: "Spawned 1 worker" },
-  { kind: "say", text: "Dispatched the changelog follow-up. Nothing else needs you." },
-];
-
-const HERO_THREADS: MockThread[] = [
-  {
-    id: "sentry",
-    title: "Triage the Sentry spike",
-    status: "running",
-    branch: "bb/triage-sentry-spike",
-    change: { files: 6, add: 124, del: 18 },
-    stream: SENTRY_STREAM,
-    transcript: [
-      { kind: "user", text: "Triage the Sentry spike on checkout." },
-      { kind: "step", text: "Explored 4 files" },
-      {
-        kind: "say",
-        text: (
-          <>
-            The spike is one error. 92% of volume: a null <code>cart</code> in{" "}
-            <code>applyPromo</code>.
-          </>
-        ),
-      },
-      { kind: "step", text: "Edited 2 files" },
-      {
-        kind: "say",
-        text: (
-          <>
-            Guarded the null case and added a regression test in{" "}
-            <code>promo.test.ts</code>. Re-running the suite.
-          </>
-        ),
-      },
-    ],
-  },
-  {
-    id: "changelog",
-    title: "Nightly changelog",
-    status: "done",
-    branch: "bb/nightly-changelog",
-    pr: 418,
-    change: { files: 1, add: 96, del: 4 },
-    transcript: [
-      { kind: "step", text: "Explored 14 commits" },
-      {
-        kind: "say",
-        text: "14 user-facing commits since yesterday. Grouped them by area.",
-      },
-      { kind: "step", text: "Edited 1 file" },
-      {
-        kind: "say",
-        text: (
-          <>
-            Wrote <code>CHANGELOG.md</code> and opened PR #418.
-          </>
-        ),
-      },
-    ],
-  },
-  {
-    id: "timeline",
-    title: "Refactor the timeline cache",
-    status: "waiting",
-    branch: "bb/timeline-cache",
-    change: { files: 3, add: 41, del: 67 },
-    transcript: [
-      {
-        kind: "user",
-        text: "Refactor the timeline cache to drop the duplicate fetch.",
-      },
-      { kind: "step", text: "Explored 3 files" },
-      { kind: "say", text: "Found the duplicate fetch. Two ways to fix it." },
-    ],
-    ask: {
-      question: "How should I dedupe the timeline fetch?",
-      options: [
-        {
-          label: "Shared in-flight promise",
-          description: "One request in flight; everyone awaits it. Simplest.",
-        },
-        {
-          label: "Short TTL cache",
-          description: "Cache the result for a few seconds, then refetch.",
-        },
-      ],
-      selected: 0,
-    },
-  },
-  {
-    id: "lin482",
-    title: "Start on LIN-482",
-    status: "running",
-    branch: "bb/lin-482-debounce-search",
-    change: { files: 2, add: 33, del: 5 },
-    stream: LIN482_STREAM,
-    transcript: [
-      { kind: "step", text: "Read LIN-482" },
-      {
-        kind: "say",
-        text: (
-          <>
-            “Debounce the search input.” Adding a 200ms debounce in{" "}
-            <code>SearchBar</code>.
-          </>
-        ),
-      },
-      { kind: "step", text: "Edited 1 file" },
-      { kind: "say", text: "Added the debounce and a test. Verifying." },
-    ],
-  },
-];
-
-// The pinned dispatcher thread, kept out of "All Threads".
-const CHIEF: MockThread = {
-  id: "chief",
-  title: "Chief",
-  status: "running",
-  branch: "bb/chief",
-  change: { files: 1, add: 12, del: 0 },
-  stream: CHIEF_STREAM,
-  transcript: [
-    { kind: "user", text: "Anything need me?" },
-    { kind: "step", text: "Swept 4 active threads" },
-    {
-      kind: "say",
-      text: (
-        <>
-          One thread is waiting on you: <code>Refactor the timeline cache</code>
-          . Sentry triage and LIN-482 are running; the nightly changelog merged.
-        </>
-      ),
-    },
-    { kind: "step", text: "Spawned 2 workers" },
-    {
-      kind: "say",
-      text: "I'll keep dispatching and ping you when something needs a call.",
-    },
-  ],
-};
-
-function ThreadStatus({ status }: { status: Status }) {
-  return (
-    <span className="tstatus" aria-hidden>
-      {status === "running" ? <Spinner className="trun" /> : null}
-      {status === "done" ? <CircleCheckIcon className="tdone" /> : null}
-      {status === "waiting" ? <MessageQuestionGlyph className="twait" /> : null}
-    </span>
-  );
-}
-
-// Cadence + rolling-window size for a running thread's live feed. The window is
-// generously larger than what fits, so the oldest rows are dropped well above
-// the (clipped) top edge and never cause a visible jump.
-const STREAM_INTERVAL_MS = 1600;
-const STREAM_WINDOW = 16;
-
-type FeedItem = { id: string; step: Step; live: boolean };
-
-/** The conversation pane. A running thread streams tool calls and messages in
- *  endlessly after its seed transcript; everything else renders statically.
- *  The first time a `spawn` step streams in, it calls `onSpawn` so the sidebar
- *  can add the nested child thread. Reduced-motion and no-JS render the seed. */
-function ThreadFeed({
-  thread,
-  onSpawn,
-}: {
-  thread: MockThread;
-  onSpawn: (parentId: string, child: MockThread) => void;
-}) {
-  const isLive = thread.status === "running" && (thread.stream?.length ?? 0) > 0;
-  const seedItems = useMemo<FeedItem[]>(
-    () =>
-      thread.transcript.map((step, i) => ({
-        id: `seed-${i}`,
-        step,
-        live: false,
-      })),
-    [thread.transcript],
-  );
-  // ThreadFeed is keyed by thread id, so switching threads remounts it and
-  // resets the stream — no in-effect reset needed.
-  const [items, setItems] = useState<FeedItem[]>(seedItems);
-
-  useEffect(() => {
-    if (!isLive) {
-      return;
-    }
-    if (window.matchMedia("(prefers-reduced-motion: reduce)").matches) {
-      return;
-    }
-    const pool = thread.stream ?? [];
-    let cursor = 0;
-    let serial = 0;
-    const id = window.setInterval(() => {
-      const step = pool[cursor % pool.length];
-      cursor += 1;
-      serial += 1;
-      if (step.kind === "spawn") {
-        onSpawn(thread.id, step.child);
-      }
-      setItems((prev) => {
-        const next = [...prev, { id: `live-${serial}`, step, live: true }];
-        return next.length > STREAM_WINDOW
-          ? next.slice(next.length - STREAM_WINDOW)
-          : next;
-      });
-    }, STREAM_INTERVAL_MS);
-    return () => window.clearInterval(id);
-  }, [thread.id, isLive, thread.stream, onSpawn]);
-
-  return (
-    <div className={isLive ? "feed feed-live" : "feed"}>
-      {items.map(({ id, step, live }, index) => {
-        // Live rows ease in as they arrive; seed rows keep the construct cascade.
-        const style: CSSProperties = live
-          ? { animation: "c-up 0.5s cubic-bezier(0.16, 1, 0.3, 1) both" }
-          : { animationDelay: `${0.66 + index * 0.09}s` };
-        if (step.kind === "user") {
-          return (
-            <div key={id} className="msg-user" style={style}>
-              {step.text}
-            </div>
-          );
-        }
-        if (step.kind === "step") {
-          return (
-            <div key={id} className="msg-step" style={style}>
-              <ChevronRight className="step-chev" />
-              {step.text}
-            </div>
-          );
-        }
-        if (step.kind === "spawn") {
-          return (
-            <div key={id} className="msg-step msg-spawn" style={style}>
-              <GitBranchIcon className="step-chev" />
-              {step.text}
-            </div>
-          );
-        }
-        return (
-          <div key={id} className="msg-say" style={style}>
-            {step.text}
-          </div>
-        );
-      })}
-    </div>
-  );
-}
-
-// The AskUserQuestion tool. Like the app, it REPLACES the prompt box: a
-// recessed card in the composer slot with the prompt, single-select option
-// rows, and Cancel / Submit answer actions.
-function AskQuestion({ ask }: { ask: Ask }) {
-  const [selected, setSelected] = useState(ask.selected);
-  return (
-    <div className="composer">
-      <div className="askq">
-        <div className="askq-q">{ask.question}</div>
-        <div className="askq-opts">
-          {ask.options.map((opt, i) => (
-            <button
-              key={opt.label}
-              type="button"
-              className={i === selected ? "askq-opt on" : "askq-opt"}
-              aria-pressed={i === selected}
-              onClick={() => setSelected(i)}
-            >
-              <span className="askq-radio">
-                {i === selected ? <CheckIcon className="askq-check" /> : null}
-              </span>
-              <span className="askq-text">
-                <span className="askq-label">{opt.label}</span>
-                <span className="askq-desc">{opt.description}</span>
-              </span>
-            </button>
-          ))}
-        </div>
-        <div className="askq-actions">
-          <span className="askq-cancel">Cancel</span>
-          <span className="askq-submit">Submit answer</span>
-        </div>
-      </div>
-    </div>
-  );
-}
-
-type DiffLine = { t: "ctx" | "add" | "del"; text: string };
-const DIFF_LINES: DiffLine[] = [
-  { t: "ctx", text: 'it("applies a valid promo", () => {' },
-  { t: "ctx", text: "  const cart = makeCart([item]);" },
-  { t: "del", text: '  expect(applyPromo(cart, "SAVE10"))' },
-  { t: "add", text: '  expect(applyPromo(cart, "SAVE10").total)' },
-  { t: "add", text: "    .toBeCloseTo(8.99);" },
-  { t: "ctx", text: "});" },
-  { t: "ctx", text: "" },
-  { t: "add", text: 'it("ignores a null cart", () => {' },
-  { t: "add", text: '  expect(() => applyPromo(null, "SAVE10"))' },
-  { t: "add", text: "    .not.toThrow();" },
-  { t: "add", text: "});" },
-];
-
-// The prompt box — used for follow-ups (with a thread) and the new-thread page
-// (no thread). Carries the full button set: expand, model picker, attach, mic,
-// send, plus the project / environment / branch / permission context row.
-function Composer({ thread }: { thread?: MockThread }) {
-  const isNew = !thread;
-  return (
-    <div className={isNew ? "composer composer-new" : "composer"}>
-      {thread ? (
-        <div className="pr-bar">
-          <GitMergeIcon className="pr-ic" />
-          <span className="pr-strong">
-            {thread.pr ? `PR #${thread.pr}` : "Working tree"}
-          </span>
-          <span className="pr-dim">
-            · {thread.pr ? "Merged" : "Uncommitted"} · {thread.change.files}{" "}
-            {thread.change.files === 1 ? "file" : "files"},
-          </span>
-          <span className="pr-add">+{thread.change.add}</span>
-          <span className="pr-del">-{thread.change.del}</span>
-          <ChevronDown className="pr-ic pr-chev" />
-        </div>
-      ) : null}
-      <div className="composer-box">
-        <div className="composer-top">
-          <textarea
-            className="composer-input"
-            rows={1}
-            placeholder={
-              isNew
-                ? "Ask anything. @ to mention files or folders"
-                : "Ask for a follow-up. @ to mention files, folders, or threads"
-            }
-            aria-label={isNew ? "Start a new thread" : "Message this thread"}
-          />
-          <Maximize2 className="cb-expand" />
-        </div>
-        <div className="composer-row">
-          <span className="model">
-            <ClaudeIcon className="model-ic" />
-            Opus 4.8 1M
-            <ChevronDown className="chev-sm" />
-          </span>
-          <span className="composer-actions" aria-hidden>
-            <Paperclip className="composer-clip" />
-            <MicIcon className="composer-clip" />
-            <span className="send-btn">
-              <SendIcon className="send-ic" />
-            </span>
-          </span>
-        </div>
-      </div>
-      <div className="context-row">
-        <span className="ctx">
-          <FolderIcon className="ctx-ic" />
-          <span>{isNew ? "paper-ultra-slop" : "bb"}</span>
-          <ChevronDown className="ctx-chev" />
-        </span>
-        <span className="ctx">
-          {isNew ? (
-            <LaptopGlyph className="ctx-ic" />
-          ) : (
-            <FolderGitIcon className="ctx-ic" />
-          )}
-          <span>{isNew ? "Work locally" : "Worktree"}</span>
-          <ChevronDown className="ctx-chev" />
-        </span>
-        <span className="ctx">
-          <GitBranchIcon className="ctx-ic" />
-          <span className="ctx-branch">
-            {isNew ? "Current (main)" : thread.branch}
-          </span>
-          <ChevronDown className="ctx-chev" />
-        </span>
-        <span className="ctx-perm">
-          Full Access
-          <ChevronDown className="ctx-chev" />
-        </span>
-        {thread && thread.status === "running" ? (
-          <Spinner className="ctx-spin" />
-        ) : null}
-      </div>
-    </div>
-  );
-}
-
-// The diff / secondary panel that opens on the right.
-function DiffPanel({
-  thread,
-  onClose,
-}: {
-  thread: MockThread;
-  onClose: () => void;
-}) {
-  return (
-    <aside className="diff-panel" aria-label="Changes">
-      <div className="diff-head">
-        <FileDiffIcon className="diff-ic" />
-        <span className="diff-title">Changes</span>
-        <span className="diff-stat pr-add">+{thread.change.add}</span>
-        <span className="diff-stat pr-del">-{thread.change.del}</span>
-        <button
-          type="button"
-          className="diff-close"
-          aria-label="Hide changes"
-          onClick={onClose}
-        >
-          <PanelRightIcon className="ri" />
-        </button>
-      </div>
-      <div className="diff-file">
-        <FolderGitIcon className="diff-file-ic" />
-        promo.test.ts
-      </div>
-      <div className="diff-body">
-        {DIFF_LINES.map((line, i) => (
-          <div key={i} className={`dl dl-${line.t}`}>
-            <span className="dl-sign">
-              {line.t === "add" ? "+" : line.t === "del" ? "-" : " "}
-            </span>
-            <span className="dl-text">{line.text || " "}</span>
-          </div>
-        ))}
-      </div>
-    </aside>
-  );
-}
-
-function HeroAppMock() {
-  const [activeId, setActiveId] = useState(HERO_THREADS[0].id);
-  const [view, setView] = useState<"thread" | "new">("thread");
-  const [diffOpen, setDiffOpen] = useState(false);
-  // Subagents a running thread spawns, keyed by parent id. They persist once
-  // spawned and render as nested child rows in the sidebar.
-  const [spawned, setSpawned] = useState<Record<string, MockThread[]>>({});
-  const spawnedChildren = useMemo(
-    () => Object.values(spawned).flat(),
-    [spawned],
-  );
-  const thread =
-    [CHIEF, ...HERO_THREADS, ...spawnedChildren].find(
-      (candidate) => candidate.id === activeId,
-    ) ?? HERO_THREADS[0];
-
-  const openThread = (id: string) => {
-    setActiveId(id);
-    setView("thread");
-  };
-
-  const handleSpawn = useCallback((parentId: string, child: MockThread) => {
-    setSpawned((prev) => {
-      const kids = prev[parentId] ?? [];
-      if (kids.some((existing) => existing.id === child.id)) {
-        return prev;
-      }
-      return { ...prev, [parentId]: [...kids, child] };
-    });
-  }, []);
-
-  return (
-    <section className="mockup-wrap">
-      <div
-        className="mock"
-        data-construct
-        aria-label="Interactive preview of the bb app"
-      >
-        <div className="mock-bar">
-          <div className="bar-left">
-            <span className="mock-dots" aria-hidden>
-              <i />
-              <i />
-              <i />
-            </span>
-            <span className="bar-menu" aria-hidden>
-              <PanelIcon className="ri bar-ic" />
-            </span>
-            <span className="bar-nav" aria-hidden>
-              <ChevronLeft className="ri" />
-              <ChevronRight className="ri" />
-            </span>
-          </div>
-          <div className="bar-main">
-            {view === "thread" ? (
-              <>
-                <span className="bar-title">{thread.title}</span>
-                <Ellipsis className="ri bar-kebab" />
-                <span className="bar-actions">
-                  <span className="editor-btn" aria-hidden>
-                    <img src={vscodeIcon} alt="" className="editor-ic" />
-                    <ChevronDown className="chev-xs" />
-                  </span>
-                  <span className="commit-btn" aria-hidden>
-                    Commit
-                  </span>
-                  <button
-                    type="button"
-                    className={diffOpen ? "bar-toggle active" : "bar-toggle"}
-                    aria-label={diffOpen ? "Hide changes" : "Show changes"}
-                    aria-pressed={diffOpen}
-                    onClick={() => setDiffOpen((open) => !open)}
-                  >
-                    <PanelRightIcon className="ri" />
-                  </button>
-                </span>
-              </>
-            ) : null}
-          </div>
-        </div>
-        <div className="mock-body">
-          <aside className="side">
-            <button
-              type="button"
-              className={
-                view === "new" ? "side-act active-act" : "side-act"
-              }
-              aria-pressed={view === "new"}
-              onClick={() => setView("new")}
-            >
-              <NewThreadIcon className="sa-ic" />
-              New thread
-            </button>
-            <div className="side-act">
-              <ClockIcon className="sa-ic" />
-              Automations
-            </div>
-            <div className="side-label">Pinned</div>
-            <button
-              type="button"
-              className={
-                view === "thread" && activeId === "chief"
-                  ? "trow trow-pin active"
-                  : "trow trow-pin"
-              }
-              aria-pressed={view === "thread" && activeId === "chief"}
-              onClick={() => openThread("chief")}
-            >
-              <span className="trow-title">Chief</span>
-            </button>
-            <div className="side-label">All Threads</div>
-            <ul className="threads">
-              {HERO_THREADS.map((candidate, index) => {
-                const isActive =
-                  view === "thread" && candidate.id === activeId;
-                const kids = spawned[candidate.id] ?? [];
-                return (
-                  <li
-                    key={candidate.id}
-                    style={{ animationDelay: `${0.6 + index * 0.06}s` }}
-                  >
-                    <button
-                      type="button"
-                      className={isActive ? "trow active" : "trow"}
-                      aria-pressed={isActive}
-                      onClick={() => openThread(candidate.id)}
-                    >
-                      <span className="trow-title">{candidate.title}</span>
-                      <ThreadStatus status={candidate.status} />
-                    </button>
-                    {kids.length > 0 ? (
-                      <ul className="threads thread-kids">
-                        {kids.map((kid) => {
-                          const kidActive =
-                            view === "thread" && kid.id === activeId;
-                          return (
-                            <li key={kid.id} className="kid-li">
-                              <button
-                                type="button"
-                                className={
-                                  kidActive
-                                    ? "trow trow-kid active"
-                                    : "trow trow-kid"
-                                }
-                                aria-pressed={kidActive}
-                                onClick={() => openThread(kid.id)}
-                              >
-                                <span className="trow-title">{kid.title}</span>
-                                <ThreadStatus status={kid.status} />
-                              </button>
-                            </li>
-                          );
-                        })}
-                      </ul>
-                    ) : null}
-                  </li>
-                );
-              })}
-            </ul>
-            <div className="side-foot" aria-hidden>
-              <GearIcon className="sa-ic" />
-            </div>
-          </aside>
-
-          {view === "thread" ? (
-            <div className="main">
-              <ThreadFeed
-                key={thread.id}
-                thread={thread}
-                onSpawn={handleSpawn}
-              />
-              {thread.ask ? (
-                <AskQuestion ask={thread.ask} />
-              ) : (
-                <Composer thread={thread} />
-              )}
-            </div>
-          ) : (
-            <div className="main main-new">
-              <Composer />
-            </div>
-          )}
-
-          {view === "thread" && diffOpen ? (
-            <DiffPanel thread={thread} onClose={() => setDiffOpen(false)} />
-          ) : null}
-        </div>
-      </div>
-    </section>
-  );
-}
-
-/* ── Band layout ──────────────────────────────────────────────────── */
-
-function Band({
-  title,
-  flip,
-  visual,
-  children,
-}: {
-  title: string;
-  flip?: boolean;
-  visual: ReactNode;
-  children: ReactNode;
-}) {
-  return (
-    <section className={flip ? "band band-flip" : "band"} data-reveal>
-      <div className="band-grid">
-        <div className="band-copy">
-          <h2>{title}</h2>
-          {children}
-        </div>
-        <div className="band-visual">{visual}</div>
-      </div>
-    </section>
-  );
-}
-
-/* ── Looping visual cycle ─────────────────────────────────────────── */
-
-/** Drives a looping visual: hold the current item, fade it out, then swap to the
- *  next and replay its entrance. Returns a monotonic `cycle` (use as the remount
- *  key; mod by item count for content) and whether it is currently fading out, so
- *  the outgoing content can ease away before the next appears. Inert under reduced
- *  motion — the first item just stays shown. */
-function useCycle(holdMs: number, fadeMs: number) {
-  const [cycle, setCycle] = useState(0);
-  const [leaving, setLeaving] = useState(false);
-  useEffect(() => {
-    if (window.matchMedia("(prefers-reduced-motion: reduce)").matches) {
-      return;
-    }
-    let holdTimer = 0;
-    let fadeTimer = 0;
-    const schedule = () => {
-      holdTimer = window.setTimeout(() => {
-        setLeaving(true);
-        fadeTimer = window.setTimeout(() => {
-          setCycle((c) => c + 1);
-          setLeaving(false);
-          schedule();
-        }, fadeMs);
-      }, holdMs);
-    };
-    schedule();
-    return () => {
-      window.clearTimeout(holdTimer);
-      window.clearTimeout(fadeTimer);
-    };
-  }, [holdMs, fadeMs]);
-  return { cycle, leaving };
-}
-
-/* ── Band visual: text the bot, bb spawns the thread ──────────────── */
-
-// A Telegram-style chat with the bb bot. The user texts a request; the bot acks
-// and a bb thread card appears, its status going spawning → running. The chat
-// shell and wallpaper stay put; only the messages cycle — they fade in, hold,
-// then fade out together before the conversation replays. CSS-only transitions.
-function AgentChat() {
-  const { cycle, leaving } = useCycle(6000, 600);
-  return (
-    <div className="tg" aria-label="Texting the Crunch bot, which spawns a bb thread">
-      <div className="tg-bar">
-        <ChevronLeft className="tg-back" />
-        <span className="tg-contact">
-          <span className="tg-name">Sawyer&rsquo;s Hermes</span>
-          <span className="tg-sub">bot</span>
-        </span>
-        <span className="tg-av" aria-hidden>
-          <img src={hermesAvatar} alt="" />
-        </span>
-      </div>
-      <div className="tg-feed">
-        <div className={leaving ? "tg-msgs leaving" : "tg-msgs"} key={cycle}>
-          <div className="tg-msg tg-out" style={{ animationDelay: "0.3s" }}>
-            <span className="tg-bubble">
-              spawn a thread to fix the failing CI on main
-              <span className="tg-time">9:41</span>
-            </span>
-          </div>
-          <div className="tg-msg tg-in" style={{ animationDelay: "1.4s" }}>
-            <span className="tg-bubble">
-              On it. Spawning a worker thread.
-              <span className="tg-cmd mono">bb spawn "fix CI on main"</span>
-            </span>
-          </div>
-          <div className="tg-msg tg-in" style={{ animationDelay: "2.4s" }}>
-            <div className="tg-thread">
-              <div className="tg-thread-top">
-                <img src={bbIcon} alt="" className="tg-thread-mark" />
-                <span className="tg-thread-eyebrow">Worker thread</span>
-                <span className="tg-stat" aria-hidden>
-                  <span
-                    className="tg-stat-spawn"
-                    style={{ animationDelay: "3.5s" }}
-                  >
-                    <Spinner className="tg-spin" />
-                    spawning
-                  </span>
-                  <span
-                    className="tg-stat-run"
-                    style={{ animationDelay: "3.5s" }}
-                  >
-                    <span className="tg-rdot" />
-                    running
-                  </span>
-                </span>
-              </div>
-              <div className="tg-thread-title">Fix CI on main</div>
-              <div className="tg-thread-branch mono">bb/fix-ci-on-main</div>
-            </div>
-          </div>
-        </div>
-      </div>
-      <div className="tg-input">
-        <Paperclip className="tg-attach" />
-        <span className="tg-field">Message</span>
-        <span className="tg-send" aria-hidden>
-          <PaperPlane className="tg-send-ic" />
-        </span>
-      </div>
-    </div>
-  );
-}
-
-/* ── Band 3 visual: mobile automation thread ─────────────────────── */
-
-type AutomationMessage = {
-  role: "user" | "agent" | "tool";
-  text: string;
-};
-
-type AutomationScenario = {
-  title: string;
-  prompt: string;
-  promptWidth: string;
-  branch: string;
-  messages: AutomationMessage[];
-};
-
-const AUTOMATION_SCENARIOS: AutomationScenario[] = [
-  {
-    title: "Issue triage",
-    prompt: "triage every new issue",
-    promptWidth: "144px",
-    branch: "bb/issue-triage",
-    messages: [
-      { role: "user", text: "triage every new issue" },
-      {
-        role: "agent",
-        text: "I'll set up an automation for every new issue.",
-      },
-      { role: "tool", text: "created trigger: on new issue" },
-      { role: "tool", text: "configured action: open triage thread" },
-      {
-        role: "agent",
-        text: "Done. New issues will get a local bb thread and Slack summary.",
-      },
-    ],
-  },
-  {
-    title: "Nightly docs sync",
-    prompt: "run nightly docs sync",
-    promptWidth: "137px",
-    branch: "bb/sync-docs",
-    messages: [
-      { role: "user", text: "run nightly docs sync" },
-      {
-        role: "agent",
-        text: "I'll create a local 2am automation for the docs sync.",
-      },
-      { role: "tool", text: "created schedule: 0 2 * * *" },
-      { role: "tool", text: "configured action: run docs worker" },
-      {
-        role: "agent",
-        text: "Done. Each run will open a thread and prepare the PR.",
-      },
-    ],
-  },
-  {
-    title: "Watch failing jobs",
-    prompt: "watch failing jobs",
-    promptWidth: "112px",
-    branch: "bb/fix-ci",
-    messages: [
-      { role: "user", text: "watch failing jobs" },
-      {
-        role: "agent",
-        text: "I'll set up an automation that reacts to failed CI jobs.",
-      },
-      { role: "tool", text: "created trigger: on job failed" },
-      { role: "tool", text: "configured action: inspect logs and spawn fix" },
-      {
-        role: "agent",
-        text: "Done. Failures will start a worker thread automatically.",
-      },
-    ],
-  },
-];
-
-// A phone-sized bb thread preview: a prompt types into the composer, sends, then
-// the matching automation transcript streams into the feed.
-function AutomationRun() {
-  const { cycle, leaving } = useCycle(7600, 500);
-  const run = AUTOMATION_SCENARIOS[cycle % AUTOMATION_SCENARIOS.length];
-  const promptStyle = {
-    "--automation-prompt-width": run.promptWidth,
-  } as CSSProperties;
-  return (
-    <div className="mockup-wrap mockup-wrap-automation">
-      <div
-        className="mock mock-automation-mobile"
-        aria-label="Mobile bb preview showing an automation prompt and streamed thread messages"
-      >
-        <div className="mock-bar">
-          <div className="bar-left">
-            <span className="bar-menu" aria-hidden>
-              <PanelIcon className="ri bar-ic" />
-            </span>
-          </div>
-          <div className="bar-main">
-            <span className="bar-title">{run.title}</span>
-            <span className="bar-actions">
-              <span className="commit-btn" aria-hidden>
-                Automation
-              </span>
-            </span>
-          </div>
-        </div>
-
-        <div
-          className={
-            leaving
-              ? "mock-body automation-body leaving"
-              : "mock-body automation-body"
-          }
-          key={cycle}
-        >
-          <div className="main">
-            <div className="feed feed-live automation-feed">
-              {run.messages.map((message, i) => {
-                const style = { animationDelay: `${3.2 + i * 0.68}s` };
-                if (message.role === "user") {
-                  return (
-                    <div
-                      className="msg-user automation-msg"
-                      key={`${message.role}-${message.text}`}
-                      style={style}
-                    >
-                      {message.text}
-                    </div>
-                  );
-                }
-                if (message.role === "tool") {
-                  return (
-                    <div
-                      className="msg-step automation-msg automation-tool"
-                      key={`${message.role}-${message.text}`}
-                      style={style}
-                    >
-                      <ChevronRight className="step-chev" />
-                      {message.text}
-                    </div>
-                  );
-                }
-                return (
-                  <div
-                    className="msg-say automation-msg"
-                    key={`${message.role}-${message.text}`}
-                    style={style}
-                  >
-                    {message.text}
-                  </div>
-                );
-              })}
-            </div>
-
-            <div className="composer automation-composer">
-              <div className="composer-box automation-composer-box">
-                <div className="composer-top">
-                  <span className="composer-input automation-typeahead">
-                    <span
-                      className="automation-type-text"
-                      style={promptStyle}
-                    >
-                      {run.prompt}
-                    </span>
-                    <span className="automation-caret" aria-hidden />
-                  </span>
-                  <Maximize2 className="cb-expand" />
-                </div>
-                <div className="composer-row">
-                  <span className="model">
-                    <OpenAiIcon className="model-ic" />
-                    Codex
-                    <ChevronDown className="chev-sm" />
-                  </span>
-                  <span className="composer-actions" aria-hidden>
-                    <Paperclip className="composer-clip" />
-                    <span className="send-btn automation-send">
-                      <SendIcon className="send-ic" />
-                    </span>
-                  </span>
-                </div>
-              </div>
-              <div className="context-row automation-context">
-                <span className="ctx">
-                  <GitBranchIcon className="ctx-ic" />
-                  <span className="ctx-branch">{run.branch}</span>
-                </span>
-                <Spinner className="ctx-spin" />
-              </div>
-            </div>
-          </div>
-        </div>
-      </div>
-    </div>
-  );
-}
-
-/* ── Band 5 visual: one agent spawns and manages a thread per provider ── */
-
-// A bb sidebar mock: a parent Claude thread with three worker threads nested
-// beneath it on a connector rail, one per provider. Each worker's status flips
-// running → done; the parent manages until they all land, then ships. Mirrors
-// the run-receipt pill and reveal timing — the list replays each cycle.
-function SpawnRow({
-  icon,
-  name,
-  task,
-  status,
-  at,
-  doneAt,
-  parent,
-}: {
-  icon: ReactNode;
-  name: string;
-  task: string;
-  status: string;
-  at: number;
-  doneAt: number;
-  parent?: boolean;
-}) {
-  return (
-    <div
-      className={parent ? "sb-thread sb-parent" : "sb-thread"}
-      style={{ animationDelay: `${at}s` }}
-    >
-      <span className="sb-prov" aria-hidden>
-        {icon}
-      </span>
-      <span className="sb-body">
-        <span className="sb-name">{name}</span>
-        <span className="sb-task">{task}</span>
-      </span>
-      <span className="sb-stat" aria-hidden>
-        <span className="sb-run" style={{ animationDelay: `${doneAt}s` }}>
-          <span className="sb-dot" />
-          {status}
-        </span>
-        <span className="sb-done" style={{ animationDelay: `${doneAt}s` }}>
-          <CheckIcon className="sb-check" />
-          done
-        </span>
-      </span>
-    </div>
-  );
-}
-
-function SpawnSidebar() {
-  const { cycle, leaving } = useCycle(5600, 500);
-  return (
-    <div
-      className="spawnbar"
-      aria-label="bb spawns and manages a worker thread for each provider"
-    >
-      <div className="sb-head">
-        <img src={bbIcon} alt="" className="sb-mark" />
-        <span className="sb-title">Threads</span>
-        <span className="sb-active">5 active</span>
-      </div>
-      <div className={leaving ? "sb-list leaving" : "sb-list"} key={cycle}>
-        <SpawnRow
-          parent
-          icon={<ClaudeIcon className="sb-ic" />}
-          name="Claude Code"
-          task="Ship the release"
-          status="managing"
-          at={0.1}
-          doneAt={4}
-        />
-        <div className="sb-kids">
-          <SpawnRow
-            icon={<OpenAiIcon className="sb-ic" />}
-            name="Codex"
-            task="Port module to TS"
-            status="running"
-            at={0.6}
-            doneAt={2.3}
-          />
-          <SpawnRow
-            icon={<CursorIcon className="sb-ic" />}
-            name="Cursor"
-            task="Refactor the auth flow"
-            status="running"
-            at={1}
-            doneAt={3}
-          />
-          <SpawnRow
-            icon={<PiIcon className="sb-ic" />}
-            name="Pi"
-            task="Write release notes"
-            status="running"
-            at={1.4}
-            doneAt={3.7}
-          />
-          <SpawnRow
-            icon={<OpencodeIcon className="sb-ic" />}
-            name="OpenCode"
-            task="Add integration tests"
-            status="running"
-            at={1.8}
-            doneAt={3.4}
-          />
-        </div>
-      </div>
-    </div>
-  );
-}
-
-/* ── Page ─────────────────────────────────────────────────────────── */
-
-function LandingPage() {
-  useScrollReveal();
-  useConstructMock();
-  useFitMock();
-  return (
-    <div className="wrap">
-      <nav className="nav">
-        <a className="logo" href="/">
-          <img src={bbIcon} alt="bb" width={36} height={36} />
-        </a>
-        <div className="nav-links">
-          <GitHubLink placement="nav">GitHub</GitHubLink>
-          <DownloadLink placement="nav" className="btn btn-primary btn-sm">
-            Download for macOS
-          </DownloadLink>
-        </div>
-      </nav>
-
-      <header className="hero">
-        <h1>
-          The First LDE<span className="lde-star">*</span>
-        </h1>
-        <p className="lde-expand">(Loop Development Environment)</p>
-        <p className="sub">
-          bb can control, customize, and automate itself, laying the groundwork
-          for your own software factory.
-        </p>
-
-        <InstallOptions placement="hero" />
-
-        <div className="providers">
-          <span className="label">Works with</span>
-          <ProviderChips />
-        </div>
-      </header>
-
-      <HeroAppMock />
-
-      <Band title="Anything can kick off work." flip visual={<AgentChat />}>
-        <p>
-          The same CLI your agents use is open to any program you write: a
-          shell script, a cron job, or your own Hermes Agent or OpenClaw bot in
-          Telegram, Signal, or Slack. Each can spawn a thread that&rsquo;s
-          waiting in your sidebar when you are.
-        </p>
-        <p>
-          It runs on your machine, and is waiting for you when you&rsquo;re back.
-        </p>
-      </Band>
-
-      <Band title="It runs without you." visual={<AutomationRun />}>
-        <p>
-          Schedule an automation to run an agent or a script on cron. Point one
-          at your tracker and it kicks off a thread for every new issue, or run
-          nightly docs, changelogs, error triage. All on your machine, not
-          someone else&rsquo;s cloud.
-        </p>
-      </Band>
-
-      <Band
-        title="The gang's all here"
-        flip
-        visual={<SpawnSidebar />}
-      >
-        <p>
-          Claude Code, Codex, Cursor, Pi, and OpenCode all live in bb. Give a
-          task to whichever fits, and have one agent spawn and manage another,
-          each in its own thread.
-        </p>
-        <p>
-          Each runs on your own subscription: the provider plan you already pay
-          for, billed by them, not bb.
-        </p>
-        <div className="providers">
-          <ProviderChips />
-        </div>
-      </Band>
-
-      <section className="statement" data-reveal>
-        <h2 className="sec-title">Fork it. Make it your own.</h2>
-        <p>
-          bb is MIT-licensed end to end. Fork the repo, customize the agents,
-          tools, and UI, and deploy your own build across your whole
-          organization. It still runs local-first on your machines, on the
-          provider subscriptions you already pay for.
-        </p>
-        <div className="cta-row">
-          <GitHubLink placement="local" className="btn btn-ghost">
-            View the source →
-          </GitHubLink>
-        </div>
-      </section>
-
-      <section className="closer" data-reveal>
-        <h2 className="sec-title">Put your agents to work.</h2>
-        <p>Free, open source, and local-first. Install in under a minute.</p>
-        <InstallOptions placement="closer" />
-        <div className="cta-row cta-row-secondary">
-          <GitHubLink placement="closer" className="btn btn-ghost">
-            View on GitHub
-          </GitHubLink>
-        </div>
-      </section>
-
-      <section className="subscribe" data-reveal>
-        <h2 className="subscribe-title">Stay in the loop.</h2>
-        <p>Product updates and what we&rsquo;re building next. No spam.</p>
-        <EmailSignup placement="footer" />
-      </section>
-
-      <footer className="footer">
-        <span>bb is free and open source (MIT)</span>
-        <span>
-          <GitHubLink placement="footer">GitHub</GitHubLink>
-          {" · "}
-          <DownloadLink placement="footer">Download</DownloadLink>
-        </span>
-      </footer>
-    </div>
-  );
-}
