--- apps/landing/tsconfig.json
@@ -1,14 +1,0 @@
-{
-  "extends": [
-    "@bb/tsconfig/base.json",
-    "@bb/tsconfig/typecheck-overrides.json"
-  ],
-  "compilerOptions": {
-    "baseUrl": ".",
-    "jsx": "react-jsx",
-    "module": "ESNext",
-    "moduleResolution": "bundler",
-    "rootDir": "."
-  },
-  "include": ["src"]
-}
