--- apps/landing/wrangler.jsonc
@@ -1,29 +1,0 @@
-// Cloudflare Workers static-assets deploy of the prerendered landing site.
-// The site is fully prerendered at build time (see vite.config.ts), so this
-// Worker only runs for first-party download redirects and the email-signup
-// endpoint; other requests are served directly from dist/client assets.
-// Deploy: pnpm exec turbo run build --filter=@bb/landing && pnpm --filter @bb/landing run deploy
-{
-  "$schema": "node_modules/wrangler/config-schema.json",
-  "name": "bb-landing",
-  "main": "./src/worker.ts",
-  "account_id": "7bb84c630057dafa53e2aacbe6bd094f",
-  "compatibility_date": "2026-06-11",
-  "assets": {
-    "directory": "./dist/client",
-    "binding": "ASSETS",
-    "run_worker_first": ["/download/macos*", "/api/subscribe"]
-  },
-  // Resend audience that /api/subscribe adds emails to ("bb users").
-  // Not a secret — only the API key is.
-  "vars": {
-    "RESEND_AUDIENCE_ID": "299db4ee-f3db-476d-9847-0bf4344d9c44"
-  },
-  "secrets": {
-    "required": ["LANDING_POSTHOG_KEY", "RESEND_API_KEY"]
-  },
-  "routes": [
-    { "pattern": "getbb.app", "custom_domain": true },
-    { "pattern": "www.getbb.app", "custom_domain": true }
-  ]
-}
