+++ apps/web/public/favicon-16x16-dark.png
