+++ apps/web/public/favicon-32x32-dark.png
