+++ apps/web/src/assets/hermes-avatar.jpg
