+++ apps/web/src/server/cloudflare.d.ts
@@ -2,4 +2,5 @@
 // Concrete binding shape is declared in Env (env.ts).
 declare module "cloudflare:workers" {
   export const env: Record<string, unknown>;
+  export function waitUntil(promise: Promise<unknown>): void;
 }
