+++ apps/web/vitest.config.ts
@@ -3,7 +3,7 @@ import { defineWorkspaceTestConfig } from "../../vitest.shared.js";
 export default defineWorkspaceTestConfig({
   test: {
     silent: "passed-only",
-    name: "@bb/landing",
+    name: "@bb/web",
     include: ["src/**/*.test.ts"],
     exclude: ["dist/**", "node_modules/**"],
   },
